+++ src/ralph_assets/views/base.py
@@ -6,21 +6,23 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 import logging
+import json
 
 from bob.data_table import DataTableColumn
+from bob.menu import MenuItem
 from bob.views.bulk_edit import BulkEditBase as BobBulkEditBase
 
 from django.conf import settings
-from django.core.exceptions import PermissionDenied
 from django.core.urlresolvers import reverse
 from django.db.models import Q
+from django.http import HttpResponse, HttpResponseBadRequest
 from django.utils.translation import ugettext_lazy as _
 from django.views.generic import TemplateView
 
 from ralph.ui.views.common import MenuMixin
 from ralph.account.models import Perm, ralph_permission
+from ralph_assets import VERSION, forms as assets_forms
 from ralph_assets.app import Assets as app
-from ralph_assets import forms as assets_forms
 from ralph_assets.models_assets import AssetType
 from ralph_assets.models import Asset
 from ralph_assets.forms import OfficeForm
@@ -45,8 +47,6 @@ class ACLGateway(object):
 
     @ralph_permission(perms)
     def dispatch(self, request, *args, **kwargs):
-        if not request.user.get_profile().has_perm(Perm.has_assets_access):
-            raise PermissionDenied
         return super(ACLGateway, self).dispatch(request, *args, **kwargs)
 
 
@@ -122,6 +122,99 @@ class AssetsBase(ACLGateway, MenuMixin, TemplateView):
             raise Exception("No form class named: {}".format(form_class_name))
         return form_class
 
+    def get_mainmenu_items(self):
+        mainmenu = [
+            MenuItem(
+                fugue_icon='fugue-building',
+                href=reverse('asset_search', kwargs={'mode': 'dc'}),
+                label=_('Data center'),
+                name='dc',
+            ),
+            MenuItem(
+                fugue_icon='fugue-printer',
+                href=reverse('asset_search', kwargs={'mode': 'back_office'}),
+                label=_('BackOffice'),
+                name='back_office',
+            ),
+            MenuItem(
+                fugue_icon='fugue-lifebuoy',
+                href=reverse('support_list'),
+                label=_('Supports'),
+                name='supports',
+            ),
+            MenuItem(
+                fugue_icon='fugue-user-green-female',
+                href=reverse('user_list'),
+                label=_('User list'),
+                name='user list',
+            ),
+            MenuItem(
+                fugue_icon='fugue-cheque',
+                href=reverse('licence_list'),
+                label=_('Licences'),
+                name='licences',
+            ),
+            MenuItem(
+                fugue_icon='fugue-table',
+                href=reverse('assets_reports'),
+                label=_('Reports'),
+                name='reports',
+            ),
+        ]
+        return mainmenu
+
+    def get_footer_items(self, details):
+        footer_items = []
+        if settings.BUGTRACKER_URL:
+            footer_items.append(
+                MenuItem(
+                    fugue_icon='fugue-bug',
+                    href=settings.BUGTRACKER_URL,
+                    label=_('Report a bug'),
+                    pull_right=True,
+                )
+            )
+        footer_items.append(
+            MenuItem(
+                fugue_icon='fugue-document-number',
+                href=settings.ASSETS_CHANGELOG_URL,
+                label=_(
+                    "Version {version}".format(
+                        version='.'.join((str(part) for part in VERSION)),
+                    ),
+                ),
+            )
+        )
+        if self.request.user.is_staff:
+            footer_items.append(
+                MenuItem(
+                    fugue_icon='fugue-toolbox',
+                    href='/admin',
+                    label=_('Admin'),
+                )
+            )
+        footer_items.append(
+            MenuItem(
+                fugue_icon='fugue-user',
+                href=reverse('user_preference', args=[]),
+                label=_('{user} (preference)'.format(user=self.request.user)),
+                pull_right=True,
+                view_args=[details or 'info', ''],
+                view_name='preference',
+            )
+        )
+        footer_items.append(
+            MenuItem(
+                fugue_icon='fugue-door-open-out',
+                href=settings.LOGOUT_URL,
+                label=_('logout'),
+                pull_right=True,
+                view_args=[details or 'info', ''],
+                view_name='logout',
+            )
+        )
+        return footer_items
+
 
 class DataTableColumnAssets(DataTableColumn):
     """
@@ -180,3 +273,20 @@ class SubmoduleModeMixin(object):
         if self.mode == 'dc':
             return 'search_dc'
         return 'search_back_office'
+
+
+class AjaxMixin(object):
+    def dispatch(self, request, *args, **kwargs):
+        if not request.is_ajax():
+            return HttpResponseBadRequest()
+        return super(AjaxMixin, self).dispatch(request, *args, **kwargs)
+
+
+class JsonResponseMixin(object):
+    content_type = 'application/json'
+
+    def render_json_response(self, context_data, status=200):
+        content = json.dumps(context_data)
+        return HttpResponse(
+            content, content_type=self.content_type, status=status
+        )