+++ src/ralph_assets/views/report.py
@@ -6,17 +6,24 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 import logging
+import uuid
 
+from bob import csvutil
+from bob.menu import MenuItem, MenuHeader
+from django.core.urlresolvers import reverse
 from django.db.models import Count
 from django.http import Http404
 from django.utils.translation import ugettext_lazy as _
 
+from ralph.util.reports import Report
+from ralph.discovery.models_device import Device
 from ralph_assets.views.base import AssetsBase
+from ralph_assets.others import get_assets_rows, get_licences_rows
 from ralph_assets.models_assets import (
     Asset,
     AssetModel,
-    AssetCategory,
     AssetStatus,
+    DeviceInfo,
     MODE2ASSET_TYPE,
 )
 
@@ -24,14 +31,21 @@ from ralph_assets.models_assets import (
 logger = logging.getLogger(__name__)
 
 
+def get_desc(choices_class, key, default='------'):
+    return choices_class.from_id(key) if key else default
+
+
 class ReportNode(object):
     """The basic report node. It is simple object which store name, count,
     parent and children."""
-    def __init__(self, name, count=0, parent=None, children=[], **kwargs):
+    def __init__(self, name, count=0, parent=None, children=[],
+                 link=None, **kwargs):
         self.name = name
         self.count = count
         self.parent = parent
         self.children = []
+        self.link = link
+        self.uid = uuid.uuid1()
 
     def add_child(self, child):
         self.children.append(child)
@@ -75,7 +89,7 @@ class ReportContainer(list):
             created = True
         return node, created
 
-    def add(self, name, count=0, parent=None, unique=True):
+    def add(self, name, count=0, parent=None, unique=True, link=None):
         if unique:
             new_node, created = self.get_or_create(name)
         else:
@@ -85,10 +99,11 @@ class ReportContainer(list):
         new_node.count = count
         if parent:
             if not isinstance(parent, ReportNode):
-                parent, _ = self.get_or_create(parent)
+                parent, __ = self.get_or_create(parent)
         if created:
             parent.add_child(new_node)
-        return new_node
+        new_node.link = link
+        return new_node, parent
 
     @property
     def roots(self):
@@ -110,6 +125,11 @@ class ReportContainer(list):
 
 class BaseReport(object):
     """Each report must inherit from this class."""
+    with_modes = True
+    with_counter = True
+    links = False
+    template_name = None
+
     def __init__(self):
         self.report = ReportContainer()
 
@@ -121,6 +141,9 @@ class BaseReport(object):
 
     def prepare(self, mode):
         raise NotImplemented()
+
+    def is_async(self, request):
+        return False
 
 
 class CategoryModelReport(BaseReport):
@@ -166,12 +189,12 @@ class CategoryModelStatusReport(BaseReport):
         for item in queryset:
             parent = item['model__category__name'] or 'Without category'
             name = item['model__name']
-            node = self.report.add(
+            node, __ = self.report.add(
                 name=name,
                 parent=parent,
             )
             self.report.add(
-                name=AssetStatus.from_id(item['status']),
+                name=get_desc(AssetStatus, item['status']),
                 parent=node,
                 count=item['num'],
                 unique=False
@@ -196,13 +219,13 @@ class ManufacturerCategoryModelReport(BaseReport):
 
         for item in queryset:
             manufacturer = item['manufacturer__name'] or 'Without manufacturer'
-            parent = self.report.add(
+            node, __ = self.report.add(
                 name=item['category__name'],
                 parent=manufacturer,
             )
             self.report.add(
                 name=item['name'],
-                parent=parent,
+                parent=node,
                 count=item['num'],
             )
 
@@ -225,7 +248,7 @@ class StatusModelReport(BaseReport):
             self.report.add(
                 name=item['model__name'],
                 count=item['num'],
-                parent=AssetStatus.DescFromID(item['status']),
+                parent=get_desc(AssetStatus, item['status']),
             )
 
 
@@ -296,19 +319,66 @@ class LinkedDevicesReport(BaseReport):
                 link=link,
                 unique=False,
             )
+        if root:
+            root.link = {
+                'label': 'go to search',
+                'url': '/assets/dc/search?id={}'.format(','.join(ids)),
+            }
 
-    def execute(self, mode):
-        queryset = AssetCategory.objects.filter(parent=None)
-        result = []
-        # aggr = defaultdict(int)
-        for row in queryset:
-            result.append({
-                'count': row.assetmodel_set.count(),
-                'name': row.name,
-                'extra': row.get_ancestors(),
-            })
+        device_info_ids = DeviceInfo.objects.exclude(
+            ralph_device_id=None
+        ).values_list(
+            'ralph_device_id', flat=True
+        )
+        devices = Device.objects.exclude(id__in=device_info_ids)
+        node, root = self.report.add(
+            parent=_('Devices without linked asset'),
+            name=str('Total'),
+            count=devices.count()
+        )
+        root.link = {
+            'label': 'go to search',
+            'url': '/ui/search/info/?without_asset=on'
+        }
 
-        return result
+
+class BaseRelationsReport(BaseReport):
+    template_name = 'assets/report_relations.html'
+    with_modes = True
+    links = False
+
+    def prepare(self, *args, **kwargs):
+        pass
+
+    def is_async(self, request):
+        self.export = request.GET.get('csv')
+        return self.export == 'on'
+
+    def get_response(self, request, result):
+        return csvutil.make_csv_response(data=result, filename=self.filename)
+
+
+class AssetRelationsReport(BaseRelationsReport):
+    slug = 'asset-relations'
+    name = _('Asset - relations')
+    filename = 'asset_relations.csv'
+
+    def get_result(self, *args, **kwargs):
+        filter_type = kwargs.get('mode')
+        return list(get_assets_rows(filter_type=filter_type))
+
+
+class LicenceRelationsReport(BaseRelationsReport):
+    slug = 'licence-relations'
+    name = _('Licence - relations')
+    filename = 'licence_relations.csv'
+
+    def get_result(self, *args, **kwargs):
+        filter_type = kwargs.get('mode')
+        data = []
+        for row in get_licences_rows(filter_type, True):
+            data.append([item.decode('utf-8') for item in row])
+        return data
 
 
 class ReportViewBase(AssetsBase):
@@ -318,6 +388,9 @@ class ReportViewBase(AssetsBase):
         CategoryModelStatusReport,
         ManufacturerCategoryModelReport,
         StatusModelReport,
+        LinkedDevicesReport,
+        AssetRelationsReport,
+        LicenceRelationsReport,
     ]
     modes = [
         {
@@ -334,12 +407,29 @@ class ReportViewBase(AssetsBase):
         },
     ]
 
+    def get_sidebar_items(self, base_sidebar_caption):
+        sidebar_menu = []
+        sidebar_menu += [MenuHeader(_('Reports'))]
+        sidebar_menu += [
+            MenuItem(
+                label=report.name, href=reverse('report_detail', kwargs={
+                    'mode': 'all',
+                    'slug': report.slug,
+                })
+            )
+            for report in self.reports
+        ]
+        sidebar_menu.extend(super(ReportViewBase, self).get_sidebar_items(
+            base_sidebar_caption
+        ))
+        return sidebar_menu
+
 
 class ReportsList(ReportViewBase):
     template_name = 'assets/report_list.html'
 
 
-class ReportDetail(ReportViewBase):
+class ReportDetail(Report, ReportViewBase):
     template_name = 'assets/report_detail.html'
 
     @property
@@ -351,6 +441,19 @@ class ReportDetail(ReportViewBase):
             if report.slug == slug:
                 return report()
         return None
+
+    def get_template_names(self, *args, **kwargs):
+        return [self.report.template_name or self.template_name]
+
+    def is_async(self, request, *args, **kwargs):
+        return self.report.is_async(request)
+
+    def get_result(self, request, *args, **kwargs):
+        report = self.get_report(kwargs.get('slug'))
+        return report.get_result(*args, **kwargs)
+
+    def get_response(self, request, result):
+        return self.report.get_response(request, result)
 
     def dispatch(self, request, *args, **kwargs):
         self.slug = kwargs.pop('slug')